testglobal/requirements.txt
coverage==6.3.2
Django==3.2.9
djangorestframework==3.12.4
model-bakery==1.1.0
pytest==7.0.0
pytest-cov==2.10.1
pytest-django==3.9.0
six==1.16.0